cli checkout implementation, index update, bug fix
- checkout implementation to cli (added to main.ts) and main class
- api checkout small fixes
- IEdoIndex update (to contain messages)
- comments added to CsvUtils, EndevorRestApi
-  fixes in FileUtils for incorrect if blocks
