+++ src/api/doc/IEdoIndex.ts
@@ -4,6 +4,7 @@
 export interface IEdoIndex {
 	prev: string;
 	stgn: string;
+	mesg: string;
 	type: string;
 	elem: { [key: string]: string };
 }
